refactor: move delete implementation into it's own file
