+++ rust/lance/src/dataset.rs
@@ -16,7 +16,6 @@ use lance_core::cache::LanceCache;
 use lance_core::datatypes::{OnMissing, OnTypeMismatch, Projectable, Projection};
 use lance_core::traits::DatasetTakeRows;
 use lance_core::utils::address::RowAddress;
-use lance_core::utils::tokio::get_num_compute_intensive_cpus;
 use lance_core::utils::tracing::{AUDIT_MODE_CREATE, AUDIT_TYPE_MANIFEST, TRACE_FILE_AUDIT};
 use lance_core::ROW_ADDR;
 use lance_datafusion::projection::ProjectionPlan;
@@ -1013,46 +1012,7 @@ impl Dataset {
 
     /// Delete rows based on a predicate.
     pub async fn delete(&mut self, predicate: &str) -> Result<()> {
-        let mut updated_fragments: Vec<Fragment> = Vec::new();
-        let mut deleted_fragment_ids: Vec<u64> = Vec::new();
-        stream::iter(self.get_fragments())
-            .map(|f| async move {
-                let old_fragment = f.metadata.clone();
-                let new_fragment = f.delete(predicate).await?.map(|f| f.metadata);
-                Ok((old_fragment, new_fragment))
-            })
-            .buffer_unordered(get_num_compute_intensive_cpus())
-            // Drop the fragments that were deleted.
-            .try_for_each(|(old_fragment, new_fragment)| {
-                if let Some(new_fragment) = new_fragment {
-                    if new_fragment != old_fragment {
-                        updated_fragments.push(new_fragment);
-                    }
-                } else {
-                    deleted_fragment_ids.push(old_fragment.id);
-                }
-                futures::future::ready(Ok::<_, crate::Error>(()))
-            })
-            .await?;
-
-        let transaction = Transaction::new(
-            self.manifest.version,
-            Operation::Delete {
-                updated_fragments,
-                deleted_fragment_ids,
-                predicate: predicate.to_string(),
-            },
-            // No change is needed to the blobs dataset.  The blobs are implicitly deleted since the
-            // rows that reference them are deleted.
-            /*blobs_op=*/
-            None,
-            None,
-        );
-
-        self.apply_commit(transaction, &Default::default(), &Default::default())
-            .await?;
-
-        Ok(())
+        write::delete::delete(self, predicate).await
     }
 
     pub async fn count_deleted_rows(&self) -> Result<usize> {
@@ -1914,7 +1874,7 @@ mod tests {
     use crate::dataset::transaction::DataReplacementGroup;
     use crate::dataset::WriteMode::Overwrite;
     use crate::index::vector::VectorIndexParams;
-    use crate::utils::test::{copy_test_data_to_tmp, TestDatasetGenerator};
+    use crate::utils::test::copy_test_data_to_tmp;
 
     use arrow::array::{as_struct_array, AsArray, GenericListBuilder, GenericStringBuilder};
     use arrow::compute::concat_batches;
@@ -3687,176 +3647,6 @@ mod tests {
         for i in 0..key.len() {
             assert_eq!(key.value(i) + 1, new_value.value(i));
         }
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_delete(
-        #[values(LanceFileVersion::Legacy, LanceFileVersion::Stable)]
-        data_storage_version: LanceFileVersion,
-        #[values(false, true)] with_scalar_index: bool,
-    ) {
-        use std::collections::HashSet;
-
-        fn sequence_data(range: Range<u32>) -> RecordBatch {
-            let schema = Arc::new(ArrowSchema::new(vec![
-                ArrowField::new("i", DataType::UInt32, false),
-                ArrowField::new("x", DataType::UInt32, false),
-            ]));
-            RecordBatch::try_new(
-                schema,
-                vec![
-                    Arc::new(UInt32Array::from_iter_values(range.clone())),
-                    Arc::new(UInt32Array::from_iter_values(range.map(|v| v * 2))),
-                ],
-            )
-            .unwrap()
-        }
-        // Write a dataset
-        let test_dir = tempdir().unwrap();
-        let test_uri = test_dir.path().to_str().unwrap();
-
-        let schema = Arc::new(ArrowSchema::new(vec![
-            ArrowField::new("i", DataType::UInt32, false),
-            ArrowField::new("x", DataType::UInt32, false),
-        ]));
-        let data = sequence_data(0..100);
-        // Split over two files.
-        let batches = vec![data.slice(0, 50), data.slice(50, 50)];
-        let mut dataset = TestDatasetGenerator::new(batches, data_storage_version)
-            .make_hostile(test_uri)
-            .await;
-
-        if with_scalar_index {
-            dataset
-                .create_index(
-                    &["i"],
-                    IndexType::Scalar,
-                    Some("scalar_index".to_string()),
-                    &ScalarIndexParams::default(),
-                    false,
-                )
-                .await
-                .unwrap();
-        }
-
-        // Delete nothing
-        dataset.delete("i < 0").await.unwrap();
-        dataset.validate().await.unwrap();
-
-        // We should not have any deletion file still
-        let fragments = dataset.get_fragments();
-        assert_eq!(fragments.len(), 2);
-        assert_eq!(dataset.count_fragments(), 2);
-        assert_eq!(dataset.count_deleted_rows().await.unwrap(), 0);
-        assert_eq!(dataset.manifest.max_fragment_id(), Some(1));
-        assert!(fragments[0].metadata.deletion_file.is_none());
-        assert!(fragments[1].metadata.deletion_file.is_none());
-
-        // Delete rows
-        dataset.delete("i < 10 OR i >= 90").await.unwrap();
-        dataset.validate().await.unwrap();
-
-        // Verify result:
-        // There should be a deletion file in the metadata
-        let fragments = dataset.get_fragments();
-        assert_eq!(fragments.len(), 2);
-        assert_eq!(dataset.count_fragments(), 2);
-        assert!(fragments[0].metadata.deletion_file.is_some());
-        assert!(fragments[1].metadata.deletion_file.is_some());
-        assert_eq!(
-            fragments[0]
-                .metadata
-                .deletion_file
-                .as_ref()
-                .unwrap()
-                .num_deleted_rows,
-            Some(10)
-        );
-        assert_eq!(
-            fragments[1]
-                .metadata
-                .deletion_file
-                .as_ref()
-                .unwrap()
-                .num_deleted_rows,
-            Some(10)
-        );
-
-        // The deletion file should contain 20 rows
-        assert_eq!(dataset.count_deleted_rows().await.unwrap(), 20);
-        // First fragment has 0..10 deleted
-        let deletion_vector = fragments[0].get_deletion_vector().await.unwrap().unwrap();
-        assert_eq!(deletion_vector.len(), 10);
-        assert_eq!(
-            deletion_vector.iter().collect::<HashSet<_>>(),
-            (0..10).collect::<HashSet<_>>()
-        );
-        // Second fragment has 90..100 deleted
-        let deletion_vector = fragments[1].get_deletion_vector().await.unwrap().unwrap();
-        assert_eq!(deletion_vector.len(), 10);
-        // The second fragment starts at 50, so 90..100 becomes 40..50 in local row ids.
-        assert_eq!(
-            deletion_vector.iter().collect::<HashSet<_>>(),
-            (40..50).collect::<HashSet<_>>()
-        );
-        let second_deletion_file = fragments[1].metadata.deletion_file.clone().unwrap();
-
-        // Delete more rows
-        dataset.delete("i < 20").await.unwrap();
-        dataset.validate().await.unwrap();
-
-        // Verify result
-        assert_eq!(dataset.count_deleted_rows().await.unwrap(), 30);
-        let fragments = dataset.get_fragments();
-        assert_eq!(fragments.len(), 2);
-        assert!(fragments[0].metadata.deletion_file.is_some());
-        let deletion_vector = fragments[0].get_deletion_vector().await.unwrap().unwrap();
-        assert_eq!(deletion_vector.len(), 20);
-        assert_eq!(
-            deletion_vector.iter().collect::<HashSet<_>>(),
-            (0..20).collect::<HashSet<_>>()
-        );
-        // Second deletion vector was not rewritten
-        assert_eq!(
-            fragments[1].metadata.deletion_file.as_ref().unwrap(),
-            &second_deletion_file
-        );
-
-        // Delete full fragment
-        dataset.delete("i >= 50").await.unwrap();
-        dataset.validate().await.unwrap();
-
-        // Verify second fragment is fully gone
-        let fragments = dataset.get_fragments();
-        assert_eq!(fragments.len(), 1);
-        assert_eq!(dataset.count_fragments(), 1);
-        assert_eq!(fragments[0].id(), 0);
-
-        // Verify the count_deleted_rows only contains the rows from the first fragment
-        // i.e. - deleted_rows from the fragment that has been deleted are not counted
-        assert_eq!(dataset.count_deleted_rows().await.unwrap(), 20);
-
-        // Append after delete
-        let data = sequence_data(0..100);
-        let batches = RecordBatchIterator::new(vec![data].into_iter().map(Ok), schema.clone());
-        let write_params = WriteParams {
-            mode: WriteMode::Append,
-            ..Default::default()
-        };
-        let dataset = Dataset::write(batches, test_uri, Some(write_params))
-            .await
-            .unwrap();
-
-        dataset.validate().await.unwrap();
-
-        let fragments = dataset.get_fragments();
-        assert_eq!(fragments.len(), 2);
-        assert_eq!(dataset.count_fragments(), 2);
-        // Fragment id picks up where we left off
-        assert_eq!(fragments[0].id(), 0);
-        assert_eq!(fragments[1].id(), 2);
-        assert_eq!(dataset.manifest.max_fragment_id(), Some(2));
     }
 
     #[rstest]
